feat: move spiers_cpp code inside oxspires_utils
